// src/generated/client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  EDITOR: 'EDITOR',
  WRITER: 'WRITER',
  READER: 'READER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const MagazineStatus = {
  DRAFT: 'DRAFT',
  REVIEW: 'REVIEW',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED'
} as const

export type MagazineStatus = (typeof MagazineStatus)[keyof typeof MagazineStatus]


export const PostType = {
  BLOG: 'BLOG',
  ARTICLE: 'ARTICLE',
  FEATURE: 'FEATURE',
  INTERVIEW: 'INTERVIEW'
} as const

export type PostType = (typeof PostType)[keyof typeof PostType]


export const PostStatus = {
  DRAFT: 'DRAFT',
  REVIEW: 'REVIEW',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED'
} as const

export type PostStatus = (typeof PostStatus)[keyof typeof PostStatus]


export const PostVisibility = {
  PUBLIC: 'PUBLIC',
  PRIVATE: 'PRIVATE',
  SUBSCRIBERS_ONLY: 'SUBSCRIBERS_ONLY'
} as const

export type PostVisibility = (typeof PostVisibility)[keyof typeof PostVisibility]


export const CommentStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED'
} as const

export type CommentStatus = (typeof CommentStatus)[keyof typeof CommentStatus]


export const MediaType = {
  IMAGE: 'IMAGE',
  VIDEO: 'VIDEO',
  AUDIO: 'AUDIO',
  DOCUMENT: 'DOCUMENT'
} as const

export type MediaType = (typeof MediaType)[keyof typeof MediaType]